refactor(move): update page move modal with new components
