+++ src/components/PageMoveModal/index.js
@@ -1,0 +1,1 @@
+export { PageMoveModal } from "./PageMoveModal"
